DATA_PIPELINE/superstore_pipeline/etl/requirements.txt
pandas
psycopg2-binary
dbt-postgres
matplotlib
